fix: fix conversion function litteral
The problem is that some metrics in the stats namespace are always
wrong: either unit counters are multiplied by 1e6, or millisecond
counters are NOT multiplied by 1e6. The reason is that if two metrics in
the same namespace have different values for the factor field, one of
these values is being chosen randomly in NewExporter and used for all
metrics.

That's because we range over a map and use values in a closure inside
the range loop.

The fix is to declare a variable inside the range loop.